Consolidate fireEvent import and split TierRanking spec into own file
