+++ src/app/groups/[id]/categories/[categoryId]/picks/[pickId]/TierRankingView.spec.tsx
@@ -1,8 +1,6 @@
 import { describe, it, expect, afterEach, vi } from "vitest";
-import { render, screen, cleanup } from "@testing-library/react";
-import { fireEvent } from "@testing-library/react";
+import { render, screen, cleanup, fireEvent } from "@testing-library/react";
 import { TierRankingView } from "./TierRankingView";
-import { TierRanking } from "./TierRanking";
 import { RankingTier, TIER_RANKING_COPY } from "./TierRanking.copy";
 import type { Option } from "@/lib/types/option";
 
@@ -123,72 +121,6 @@ describe("TierRankingView", () => {
       );
       fireEvent.click(screen.getByText("Matrix"));
       expect(onOptionClick).toHaveBeenCalledWith("opt-99");
-    });
-  });
-});
-
-describe("TierRanking", () => {
-  describe("cycling tier on click", () => {
-    it("starts with options in Unranked", () => {
-      const option = makeOption({ id: "opt-1", title: "Inception" });
-      render(<TierRanking options={[option]} />);
-      expect(screen.getByText("Inception")).toBeDefined();
-    });
-
-    it("advances an option to LoveIt on first click", () => {
-      const option = makeOption({ id: "opt-1", title: "Inception" });
-      render(<TierRanking options={[option]} />);
-      fireEvent.click(screen.getByText("Inception"));
-      const loveItHeader = screen.getByText(
-        TIER_RANKING_COPY.tiers[RankingTier.LoveIt],
-      );
-      expect(loveItHeader).toBeDefined();
-      expect(screen.getAllByText("Inception")).toBeDefined();
-    });
-
-    it("cycles through all tiers and back to Unranked", () => {
-      const option = makeOption({ id: "opt-1", title: "Inception" });
-      render(<TierRanking options={[option]} />);
-
-      const getChip = () => screen.getByText("Inception");
-      const getTierSectionFor = (chip: HTMLElement) => {
-        // Walk up to the tier bucket div (the one with the h3 sibling)
-        let el: HTMLElement | null = chip;
-        while (el && el.previousElementSibling?.tagName !== "H3") {
-          el = el.parentElement;
-        }
-        return el?.previousElementSibling?.textContent ?? null;
-      };
-
-      // Unranked → LoveIt
-      fireEvent.click(getChip());
-      expect(getTierSectionFor(getChip())).toBe(
-        TIER_RANKING_COPY.tiers[RankingTier.LoveIt],
-      );
-
-      // LoveIt → Yes
-      fireEvent.click(getChip());
-      expect(getTierSectionFor(getChip())).toBe(
-        TIER_RANKING_COPY.tiers[RankingTier.Yes],
-      );
-
-      // Yes → Maybe
-      fireEvent.click(getChip());
-      expect(getTierSectionFor(getChip())).toBe(
-        TIER_RANKING_COPY.tiers[RankingTier.Maybe],
-      );
-
-      // Maybe → NotReally
-      fireEvent.click(getChip());
-      expect(getTierSectionFor(getChip())).toBe(
-        TIER_RANKING_COPY.tiers[RankingTier.NotReally],
-      );
-
-      // NotReally → Unranked
-      fireEvent.click(getChip());
-      expect(getTierSectionFor(getChip())).toBe(
-        TIER_RANKING_COPY.tiers[RankingTier.Unranked],
-      );
     });
   });
 });
